backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
pymongo==4.10.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
bcrypt==4.2.1
requests==2.32.3
openai==1.58.1
reportlab==4.2.5

